backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary>=2.9.10

# Validation and security
pydantic[email]==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# External services (Stripe, Anthropic)
stripe==11.3.0
anthropic==0.43.0

# HTTP requests
httpx==0.28.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
